backend/requirements.txt
# KUCCPS Career Hub Backend Dependencies

Flask==2.3.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
psycopg2-binary==2.9.6
requests==2.31.0
Werkzeug==2.3.0
click==8.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
SQLAlchemy==2.0.19
reportlab==4.0.4
PyPDF2==3.0.1
twilio==8.9.0
beautifulsoup4==4.12.2
lxml==4.9.3
